Generalize data generation
