+++ package.json
@@ -17,13 +17,13 @@
     "publish-next": "lerna publish --npm-tag=next --bump=prerelease",
     "start:ec": "start-storybook -p 9001 -c src/ec/.storybook",
     "start:eu": "start-storybook -p 9001 -c src/eu/.storybook",
-    "render": "npx lerna exec --scope @ecl-twig/php-renderer -- yarn run render",
+    "render:dataGen": "babel-node ./scripts/generate-data-files.js",
     "render:setup": "npx lerna exec --scope @ecl-twig/php-renderer -- yarn run setup",
-    "render:clean": "npx lerna exec --scope @ecl-twig/php-renderer -- yarn run clean",
     "test": "npm run jest"
   },
   "devDependencies": {
-    "@babel/core": "7.3.4",
+    "@babel/core": "7.4.0",
+    "@babel/node": "7.2.2",
     "@babel/preset-env": "7.3.4",
     "@babel/preset-react": "7.0.0",
     "@ecl-twig/storybook-addon-code": "file:utils/storybook-addon-code",
@@ -46,6 +46,7 @@
     "eslint-plugin-prettier": "3.0.1",
     "eslint-plugin-react": "7.12.4",
     "eslint-plugin-unicorn": "7.1.0",
+    "fs-extra": "7.0.1",
     "husky": "1.3.1",
     "jest": "24.3.1",
     "js-beautify": "1.9.0",
