+++ src/ec/packages/ec-components/package.json
@@ -8,6 +8,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@ecl-twig/ec-component-accordion": "^0.0.1-alpha",
     "@ecl-twig/ec-component-blockquote": "^0.0.1-alpha",
     "@ecl-twig/ec-component-breadcrumb": "^0.0.1-alpha",
     "@ecl-twig/ec-component-button": "^0.0.1-alpha",
