--- src/php/packages/renderer/ec/blockquote/blockquote.html
@@ -1,9 +1,0 @@
-<blockquote class="ecl-blockquote"
-  ><p class="ecl-blockquote__body"
-    >An interconnected grid will help deliver the ultimate goal of the Energy
-    Union, to ensure affordable, secure and sustainable energy, and also growth
-    across the EU.</p
-  ><footer class="ecl-blockquote__attribution"
-    ><cite class="ecl-blockquote__author">President Juncker</cite></footer
-  ></blockquote
->
