--- src/php/packages/renderer/ec/blockquote/data.json
@@ -1,4 +1,0 @@
-{
-  "citation": "An interconnected grid will help deliver the ultimate goal of the Energy Union, to ensure affordable, secure and sustainable energy, and also growth across the EU.",
-  "author": "President Juncker"
-}
