--- src/php/packages/renderer/ec/breadcrumb/breadcrumb-simple.html
@@ -1,3 +1,0 @@
-<nav class="ecl-breadcrumb" aria-label="" data-ecl-breadcrumb="true"
-  ><ol class="ecl-breadcrumb__container"></ol
-></nav>
