--- src/php/packages/renderer/ec/breadcrumb/data-simple.json
@@ -1,8 +1,0 @@
-{
-  "items": [
-    { "label": "Home", "href": "/example" },
-    { "label": "About the European Commission", "href": "/example" },
-    { "label": "News" }
-  ],
-  "label": "You are here:"
-}
