--- src/php/packages/renderer/ec/breadcrumb/data.json
@@ -1,10 +1,0 @@
-{
-  "items": [
-    { "label": "Home", "href": "/example" },
-    { "label": "About the European Commission", "href": "/example" },
-    { "label": "Organisational structure", "href": "/example" },
-    { "label": "How the Commission is organised", "href": "/example" },
-    { "label": "News" }
-  ],
-  "label": "You are here:"
-}
