--- src/php/packages/renderer/ec/button/button--call.html
@@ -1,7 +1,0 @@
-<button class="ecl-button ecl-button--call" type="submit"
-  ><span class="ecl-button__container"
-    ><span class="ecl-button__label" data-ecl-label
-      >Call to action button</span
-    ></span
-  ></button
->
