--- src/php/packages/renderer/ec/button/button--ghost.html
@@ -1,5 +1,0 @@
-<button class="ecl-button ecl-button--ghost" type="submit"
-  ><span class="ecl-button__container"
-    ><span class="ecl-button__label" data-ecl-label>Ghost button</span></span
-  ></button
->
