--- src/php/packages/renderer/ec/button/button--primary.html
@@ -1,5 +1,0 @@
-<button class="ecl-button ecl-button--primary" type="submit"
-  ><span class="ecl-button__container"
-    ><span class="ecl-button__label" data-ecl-label>Primary button</span></span
-  ></button
->
