--- src/php/packages/renderer/ec/button/button--search.html
@@ -1,5 +1,0 @@
-<button class="ecl-button ecl-button--search" type="submit"
-  ><span class="ecl-button__container"
-    ><span class="ecl-button__label" data-ecl-label>Search button</span></span
-  ></button
->
