--- src/php/packages/renderer/ec/button/button--secondary.html
@@ -1,7 +1,0 @@
-<button class="ecl-button ecl-button--secondary" type="submit"
-  ><span class="ecl-button__container"
-    ><span class="ecl-button__label" data-ecl-label
-      >Secondary button</span
-    ></span
-  ></button
->
