--- src/php/packages/renderer/ec/button/data--call.json
@@ -1,9 +1,0 @@
-{
-  "label": "Call to action button",
-  "icon": {
-    "shape": "ui--corner-arrow",
-    "size": "xs",
-    "transform": "rotate-90"
-  },
-  "variant": "call"
-}
