--- src/php/packages/renderer/ec/button/data--ghost.json
@@ -1,1 +1,0 @@
-{ "label": "Ghost button", "variant": "ghost" }
