--- src/php/packages/renderer/ec/button/data--primary.json
@@ -1,1 +1,0 @@
-{ "label": "Primary button", "variant": "primary" }
