--- src/php/packages/renderer/ec/button/data--search.json
@@ -1,1 +1,0 @@
-{ "label": "Search button", "variant": "search" }
