--- src/php/packages/renderer/ec/button/data--secondary.json
@@ -1,1 +1,0 @@
-{ "label": "Secondary button", "variant": "secondary" }
