--- src/php/packages/renderer/ec/card/data--card.json
@@ -1,26 +1,0 @@
-{
-  "image": { "alt": "card image", "src": "/example-image.jpg" },
-  "meta": "Meta1 | Meta2",
-  "title": {
-    "variant": "standalone",
-    "label": "Better regulation",
-    "href": "/example",
-    "level": 1
-  },
-  "description": "Transparently designing and evaluating evidence-based EU legislation, backed by citizens views.",
-  "infos": [
-    {
-      "icon": { "shape": "general--calendar", "size": "xs" },
-      "label": "2018/10/22"
-    },
-    {
-      "icon": { "shape": "general--location", "size": "xs" },
-      "label": "Luxembourg"
-    }
-  ],
-  "tags": [
-    { "label": "tag 1", "href": "/example" },
-    { "label": "tag 2", "href": "/example" },
-    { "label": "tag 3", "href": "/example" }
-  ]
-}
