--- src/php/packages/renderer/ec/card/data--tile.json
@@ -1,13 +1,0 @@
-{
-  "title": {
-    "variant": "standalone",
-    "label": "Better regulation",
-    "level": 1
-  },
-  "description": "Transparently designing and evaluating evidence-based EU legislation, backed by citizens views.",
-  "links": [
-    { "label": "link 1", "href": "/example", "variant": "standalone" },
-    { "label": "link 2", "href": "/example", "variant": "standalone" },
-    { "label": "link 3", "href": "/example", "variant": "standalone" }
-  ]
-}
